refactor(crates/server): refactor `backend`
